backend/pyproject.toml
[project]
name = "atomix-backend"
version = "0.1.0"
description = "ATOMIX backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "docker>=7.1.0",
    "dotenv>=0.9.9",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
